implemented user registration

+++ library/authentication.test.ts
@@ -1,8 +1,9 @@
-import { afterAll, describe, expect, it } from "vitest";
-import { getSignedInUser, isUserSignedIn } from "./authentication.ts";
+import { afterEach, describe, expect, it } from "vitest";
+import { createNewAccount, getSignedInUser, isUserSignedIn, signInUser } from "./authentication.ts";
 import { v4 as uuidv4 } from "uuid";
 import { query } from "../database/adapter.js";
 import type { FastifyRequest } from "fastify";
+import * as argon2 from "argon2";
 
 const passwordHash =
 	"$argon2id$v=19$m=65536,t=3,p=4$P5wGfnyG6tNP2iwvWPp9SA$Gp3wgJZC1xe6fVzUTMmqgCGgFPyZeCt1aXjUtlwSMmo";
@@ -11,13 +12,17 @@ describe("authentication", () => {
 	const userIds: string[] = [];
 	const sessionIds: string[] = [];
 
-	afterAll(async function () {
-		await query(
-			`DELETE FROM sessions WHERE id IN (${sessionIds.map((sessionId) => `'${sessionId}'`).join(", ")})`,
-		);
-		await query(
-			`DELETE FROM users WHERE id IN (${userIds.map((userId) => `'${userId}'`).join(", ")})`,
-		);
+	afterEach(async function () {
+		if (sessionIds.length) {
+			await query(
+				`DELETE FROM sessions WHERE id IN (${sessionIds.map((sessionId) => `'${sessionId}'`).join(", ")})`,
+			);
+		}
+		if (userIds.length) {
+			await query(
+				`DELETE FROM users WHERE id IN (${userIds.map((userId) => `'${userId}'`).join(", ")})`,
+			);
+		}
 	});
 
 	it("isUserSignedIn should return false if session is empty", async () => {
@@ -88,6 +93,54 @@ describe("authentication", () => {
 			name: "Dylan",
 			surname: "George",
 			username: "DylanG",
+		});
+	});
+
+	it("createNewAccount should create new account and return user id", async () => {
+		const userId = await createNewAccount({
+			name: "Seth",
+			surname: "Milchick",
+			password: "a test",
+			username: "sMilchick",
+		});
+		userIds.push(userId);
+
+		const newUserResult = await query("SELECT * FROM users WHERE id = $1", [userId]);
+
+		expect(newUserResult.rowCount).toEqual(1);
+		expect(newUserResult.rows[0].firstname).toEqual("Seth");
+		expect(newUserResult.rows[0].lastname).toEqual("Milchick");
+		expect(newUserResult.rows[0].username).toEqual("sMilchick");
+		const passwordMatches = await argon2.verify(newUserResult.rows[0].password, "a test");
+		expect(passwordMatches).toStrictEqual(true);
+	});
+
+	it("signInUser should create a new session for provided user id and session id", async () => {
+		let cookieHandlerArgs: unknown[] = [];
+		const userId = await createNewAccount({
+			name: "Harmony",
+			surname: "Cobel",
+			password: "a test",
+			username: "hCobel",
+		});
+		const sessionId = await signInUser(userId, (...args) => {
+			cookieHandlerArgs = args;
+		});
+		userIds.push(userId);
+		sessionIds.push(sessionId);
+
+		const userSessionResult = await query("SELECT * FROM sessions WHERE user_id = $1", [userId]);
+
+		expect(userSessionResult.rowCount).toEqual(1);
+		expect(userSessionResult.rows[0].id).toEqual(sessionId);
+		expect(userSessionResult.rows[0].user_id).toEqual(userId);
+		expect(cookieHandlerArgs[0]).toEqual("session");
+		expect(cookieHandlerArgs[1]).toEqual(sessionId);
+		expect(cookieHandlerArgs[2]).toEqual({
+			httpOnly: true,
+			maxAge: 604800,
+			sameSite: "strict",
+			secure: "auto",
 		});
 	});
 });
